+++ src/main/java/net/tryplay/command.java
@@ -1,4 +1,4 @@
-package com.tryplay.net.com.tryplay.net.Com.tryplay;
+package net.tryplay;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
